requirements.txt
python-telegram-bot[ext]==20.5
python-dotenv==1.0.0
yt-dlp==2023.10.7
ffmpeg==7.1
